Moved internals to separate workspace
