+++ src/main/scala/internals/Output.scala
@@ -1,0 +1,17 @@
+package internals
+
+import org.apache.spark.rdd.RDD
+
+sealed trait Output
+
+class DatasetOutput(getRDD: => RDD[_]) extends Output {
+  lazy val get: RDD[_] = getRDD
+}
+
+class DatumOutput(getDatum: => Any) extends Output {
+  lazy val get: Any = getDatum
+}
+
+class TransformerOutput(getTransformer: => TransformerOperator) extends Output {
+  lazy val get: TransformerOperator = getTransformer
+}
